modified:   conf/airframes/AGR/quad_680-ahrs_float_euler-stab_float_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_float_mlkf-stab_float_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_float_mlkf-stab_int_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_float_quat-stab-int_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_float_quat-stab_float_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_int_euler-stab-int_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_int_euler-stab_float_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_int_quat-stab_float_euler.xml 	modified:   conf/airframes/AGR/quad_680-ahrs_int_quat-stab_int_euler.xml 	modified:   conf/airframes/examples/quadrotor_navgo.xml 	modified:   conf/flight_plans/AGR/coca_cola.xml 	modified:   conf/modules/nav_tower_rotorcraft.xml 	modified:   sw/airborne/modules/nav/nav_tower_rotorcraft.c 	modified:   sw/airborne/modules/nav/nav_tower_rotorcraft.h

+++ sw/airborne/modules/nav/nav_tower_rotorcraft.h
@@ -32,7 +32,7 @@
 #define NAV_TOWER_ROTORCRAFT_H
 
 #include "firmwares/rotorcraft/navigation.h"
-extern int32_t carrot1;
+extern float carrot1;
 extern int8_t sign_radius;
 extern bool_t nav_tower_run(struct EnuCoor_i *wp_center, int32_t radius);
